backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
grpcio==1.59.3
grpcio-tools==1.59.3
aiofiles==23.2.1
websockets==12.0
qrcode[pil]==7.4.2
cryptography==41.0.7
APScheduler==3.10.4
curl==0.0.2

